feat: VL-8606 mockOrganization to private repo
