+++ packages/components-organizations-registrar/src/index.jsx
@@ -21,4 +21,5 @@ export * from './PublicAppRoot.jsx';
 export * from './utils/auth/AuthContext';
 export * from './utils/auth/authScopes';
 export * from './utils/ConfigContext';
+export * from './utils/chainNames';
 export * from './state';
