+++ packages/components-organizations-registrar/src/pages/organizations/utils/index.js
@@ -20,25 +20,3 @@ export const requestOptions = {
 
 export const organizationPlaceholder =
   'Add a few words describing your organization (boilerplate text) so that other Network participants can learn about it.';
-
-export const initialRecordMock = {
-  profile: {
-    name: `Mock Organization, Inc. ${new Date().getTime()}`,
-    website: `https://mockcompany${new Date().getTime()}.com`,
-    linkedInProfile: 'https://www.linkedin.com/company/mockcompanyid',
-    physicalAddress: { line1: '123 Mock Street' },
-    location: { countryCode: 'US' },
-    contactEmail: 'support@mockcompany.com',
-    technicalEmail: 'tech@mockcompany.com',
-    description: 'This is a mock description for local development.',
-    registrationNumbers: [{ authority: 'DunnAndBradstreet', number: '12345' }],
-    adminGivenName: 'AdminFirst',
-    adminFamilyName: 'AdminLast',
-    adminTitle: 'Administrator',
-    adminEmail: 'admin@mockcompany.com',
-    signatoryGivenName: 'SignerFirst',
-    signatoryFamilyName: 'SignerLast',
-    signatoryTitle: 'Signatory Officer',
-    signatoryEmail: 'sign@mockcompany.com',
-  },
-};
