+++ packages/components-organizations-registrar/src/pages/services/utils/isAddButtonDisabled.js
@@ -1,0 +1,13 @@
+export const isAddButtonDisabled = (
+  inProgress,
+  isIssueOrInspection,
+  selectedCAO,
+  serviceEndpoint,
+) => {
+  return (
+    !(
+      (isIssueOrInspection && serviceEndpoint && selectedCAO) ||
+      (!isIssueOrInspection && serviceEndpoint)
+    ) || inProgress
+  );
+};
